V1.1.1 wip (#2)
v1.2.0

+++ ubuntu.sh
@@ -2,6 +2,8 @@
 #set -x
 set -euo pipefail
 
+installed=""
+not_installed=""
 install_chrome() 
 {
     echo Installing CHROME
@@ -9,6 +11,7 @@ install_chrome()
     sudo sh -c 'echo "deb https://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list'
     sudo apt-get update -q 
     sudo apt-get install -y -q google-chrome-stable
+   installed+="Chrome\n"
 }
 
 set_dark_theme()
@@ -23,28 +26,41 @@ set_dark_theme()
     then
         gsettings set org.gnome.desktop.interface gtk-theme Yaru-dark # Legacy apps, can specify an accent such as Yaru-olive-dark
         gsettings set org.gnome.desktop.interface color-scheme prefer-dark # new apps
+        installed+="Dark Theme\n"
+
     else
         echo "Cannot setup theme for Ubuntu ${release}!"
+        not_installed+="Dark Theme\n"
     fi
 }
 
 install_oh_my_zsh()
 {
-    echo "Installing oh-my-zsh"
-    wget --no-cache -O "$HOME/.zshrc" "https://raw.githubusercontent.com/manuelgustavo/ubuntu-setup/main/.zshrc"
-    sudo apt-get install -y -q zsh fonts-powerline
-    # chsh -s "$(which zsh)"
-    #sudo apt-get install -y -q zsh-autosuggestions zsh-syntax-highlighting
-    # install oh-my-zsh
-    rm -fr "$HOME/.oh-my-zsh"
-    sh -c "$(wget --no-cache -O- https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh) --unattended --skip-chsh --keep-zshrc"
-    git clone --quiet --depth 1 https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
-    git clone --quiet --depth 1 https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-    
-    # Change the default shell
-    sudo sed -i -E "s/($USER.*)(bash)/\1zsh/" /etc/passwd
-    sudo update-passwd
-
+    if [[ ! -d "$HOME/.oh-my-zsh" ]]
+    then
+    {
+        echo "Installing oh-my-zsh"
+        wget --no-cache -O "$HOME/.zshrc" "https://raw.githubusercontent.com/manuelgustavo/ubuntu-setup/main/.zshrc"
+        sudo apt-get install -y -q zsh fonts-powerline
+        # chsh -s "$(which zsh)"
+        #sudo apt-get install -y -q zsh-autosuggestions zsh-syntax-highlighting
+        # install oh-my-zsh
+        rm -fr "$HOME/.oh-my-zsh"
+        sh -c "$(wget --no-cache -O- https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh) --unattended --skip-chsh --keep-zshrc"
+        git clone --quiet --depth 1 https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+        git clone --quiet --depth 1 https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+        
+        # Change the default shell
+        sudo sed -i -E "s/($USER.*)(bash)/\1zsh/" /etc/passwd
+        sudo update-passwd
+       installed+="oh-my-zsh\n"
+    }
+    else
+    {
+        echo "If you want to install oh-my-zsh, delete the ~/.oh-my-zsh directory"
+       not_installed+="oh-my-zsh\n"
+    }
+    fi
     #     wget --no-cache https://github.com/robbyrussell/oh-my-zsh/raw/master/tools/install.sh
     # sed -i.tmp 's:env zsh::g' install.sh
     # sed -i.tmp 's:chsh -s .*$::g' install.sh
@@ -62,12 +78,26 @@ install_gnome_extensions()
             gir1.2-clutter-1.0 \
             gnome-system-monitor
 
-    echo "Installing Gnome Extension -- Dash to Panel https://extensions.gnome.org/extension/1160/dash-to-panel/"
-    gdbus call --session \
-            --dest org.gnome.Shell.Extensions \
-            --object-path /org/gnome/Shell/Extensions \
-            --method org.gnome.Shell.Extensions.InstallRemoteExtension \
-            "dash-to-panel@jderose9.github.com" 2>/dev/null || true
+    declare gnome_extension
+
+    gnome_extension="dash-to-panel@jderose9.github.com"
+    if [[ ! -d "$HOME/.local/share/gnome-shell/extensions/${gnome_extension}" ]]
+    then
+    {
+        echo "Installing Gnome Extension -- Dash to Panel https://extensions.gnome.org/extension/1160/dash-to-panel/"
+        gdbus call --session \
+                --dest org.gnome.Shell.Extensions \
+                --object-path /org/gnome/Shell/Extensions \
+                --method org.gnome.Shell.Extensions.InstallRemoteExtension \
+                "${gnome_extension}" 2>/dev/null || true
+       installed+="Gnome Extension -- Dash to Panel\n"
+    }
+    else
+    {
+        echo "Skipping Dash to Panel -- already installed."
+       not_installed+="Gnome Extension -- Dash to Panel\n"
+    }
+    fi
 
     # TODO: The below needs to be re-enabled when made available!
     # echo "Installing Gnome Extension -- system-monitor https://extensions.gnome.org/extension/120/system-monitor/"
@@ -77,25 +107,49 @@ install_gnome_extensions()
     #            --method org.gnome.Shell.Extensions.InstallRemoteExtension \
     #            "system-monitor@paradoxxx.zero.gmail.com" 2>/dev/null
 
-    echo "Installing Gnome Extension -- system-monitor-next https://extensions.gnome.org/extension/3010/system-monitor-next/"
-    gdbus call --session \
+    gnome_extension="system-monitor-next@paradoxxx.zero.gmail.com"
+    if [[ ! -d "$HOME/.local/share/gnome-shell/extensions/${gnome_extension}" ]]
+    then
+    {
+        echo "Installing Gnome Extension -- system-monitor-next https://extensions.gnome.org/extension/3010/system-monitor-next/"
+        gdbus call --session \
             --dest org.gnome.Shell.Extensions \
             --object-path /org/gnome/Shell/Extensions \
             --method org.gnome.Shell.Extensions.InstallRemoteExtension \
-            "system-monitor-next@paradoxxx.zero.gmail.com" 2>/dev/null || true
+            "${gnome_extension}" 2>/dev/null || true
+       installed+="Gnome Extension -- system-monitor-next\n"
+    }
+    else
+    {
+        echo "Skipping system-monitor-next -- already installed."
+       not_installed+="Gnome Extension -- system-monitor-next\n"
+    }
+    fi
 
-    echo "Installing Gnome Extension -- Removable Drive Menu https://extensions.gnome.org/extension/7/removable-drive-menu/"
-    gdbus call --session \
+    gnome_extension="drive-menu@gnome-shell-extensions.gcampax.github.com"
+    if [[ ! -d "$HOME/.local/share/gnome-shell/extensions/${gnome_extension}" ]]
+    then
+    {
+        echo "Installing Gnome Extension -- Removable Drive Menu https://extensions.gnome.org/extension/7/removable-drive-menu/"
+        gdbus call --session \
             --dest org.gnome.Shell.Extensions \
             --object-path /org/gnome/Shell/Extensions \
             --method org.gnome.Shell.Extensions.InstallRemoteExtension \
-            "drive-menu@gnome-shell-extensions.gcampax.github.com" 2>/dev/null || true
+            "${gnome_extension}" 2>/dev/null || true
+       installed+="Gnome Extension -- Removable Drive\n"
+    }
+    else
+    {
+        echo "Skipping Removable Drive Menu -- already installed."
+       not_installed+="Gnome Extension -- Removable Drive Menu\n"
+    }
+    fi
 }
 
 install_tilix()
 {
     sudo apt-get -y -q install tilix
-    wget --no-cache -O- "https://raw.githubusercontent.com/manuelgustavo/ubuntu-setup/main/tilix.conf" | dconf load /com/gexperts/Tilix/
+    wget --no-cache -O- "https://raw.githubusercontent.com/manuelgustavo/ubuntu-setup/main/tilix_rosipov.conf" | dconf load /com/gexperts/Tilix/
     # dconf dump /com/gexperts/Tilix/
     # Install Powerline Droid Sans Mono Dotted.
     mkdir -p "$HOME/.local/share/fonts"
@@ -103,7 +157,7 @@ install_tilix()
     fc-cache -f
     sudo update-alternatives --set x-terminal-emulator /usr/bin/tilix.wrapper
     
-    sudo ln -s /etc/profile.d/vte-2.91.sh /etc/profile.d/vte.sh
+    sudo ln -s /etc/profile.d/vte-2.91.sh /etc/profile.d/vte.sh || true
     
     { 
         echo
@@ -111,6 +165,7 @@ install_tilix()
         echo '    source /etc/profile.d/vte.sh'
         echo 'fi'
     } >> "$HOME/.zshrc"
+   installed+="Tilix\n"
 }
 
 install_vscode()
@@ -130,6 +185,7 @@ install_vscode()
     sudo apt install code
     echo "Installing VSCode extensions..."
     sh -c "$(wget --no-cache -O- https://raw.githubusercontent.com/manuelgustavo/vscode-extensions/main/vscode-extensions.sh)"
+   installed+="VScode + extensions\n"
 }
 
 main()
@@ -147,7 +203,27 @@ main()
     echo .
     echo .
     echo .
-    echo "INSTALLATION SUCCESSFUL!"
+    echo "SCRIPT SUCCESS!"
+    echo .
+    echo "-------------------- Summary --------------------"
+    if [[ -n "${installed}" ]]
+    then
+    {
+        echo "Installed:"
+        printf "${installed}"
+    }
+    fi
+    echo "-------------------------------------------------"
+    if [[ -n "${not_installed}" ]]
+    then
+    {
+        echo "Skipped:"
+        printf "${not_installed}"
+    }
+    fi
+    echo "-------------------------------------------------"
+    echo .
+    echo .
     echo "It's recommended to log-off and log-on again!"
 }
